+++ ports/bvi/files/usr/local/share/bvi/plugins/colorscheme.lua
@@ -1,0 +1,10 @@
+return {
+  -- Gruvbox for unified BAUXBSD theming (lite mode)
+  { "ellisonleao/gruvbox.nvim" },
+  {
+    "LazyVim/LazyVim",
+    opts = {
+      colorscheme = "gruvbox",
+    },
+  },
+}